feat: split up functions for multiple listings and single listing, to avoid conflict
